server/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - PostgreSQL (psycopg2-binary only, no SQLAlchemy)
psycopg2-binary==2.9.9

# Configuration
pydantic==2.4.2
pydantic-settings==2.0.3

# HTTP client (for future LLM integration)
httpx==0.25.1

# Utilities
python-dotenv==1.0.0
